refactor(output): extract field tests to sibling file
